action_flows/pyproject.toml
[tool.poetry]
name = "false"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
fastapi = "^0.115.6"
python = "^3.12"
uvicorn ="^0.18.3"
pydantic_core ="^2.6.3"
langchain ="^0.3.13"
langchain-community ="^0.3.0"
pymupdf = "^1.25.1"
langchain_experimental = "^0.3.4"

langgraph ="^0.2.59"
langchain-groq ="^0.2.2"
pandas = "^2.2.3"
tabulate = "^0.9.0"
psycopg2-binary = "^2.9.10"
pydantic-settings = "^2.7.0"
sqlmodel = "^0.0.22"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
